Tune LotteryDemoContract.

+++ chain/contract/AElf.Contracts.LotteryDemoContract/LotteryDemoContractConstants.cs
@@ -2,9 +2,9 @@ namespace AElf.Contracts.LotteryDemoContract
 {
     public partial class LotteryDemoContract
     {
-        private const int DefaultPrice = 1;
+        private const long DefaultPrice = 1;
         private const long DefaultDrawingLag = 80;
-        private const int MaximumBuyAmount = 100;
-        private const int MaximumReturnAmount = 20;
+        private const long MaximumBuyAmount = 100;
+        private const long MaximumReturnAmount = 20;
     }
 }
